Updated CMake configuration to require Fortran 3.10, added include directory for logging headers, and refined library linking for different platforms. Adjusted numeric values in JSON configuration files and improved error handling in Python scripts for missing resources.

+++ Python/core/meshmagick/mesh.py
@@ -161,7 +161,7 @@ class Plane(object):
 
     def __init__(self, normal=(0., 0., 1.), scalar=0., name=None):
 
-        normal = np.asarray(normal, dtype=np.float)
+        normal = np.asarray(normal, dtype=np.float64)
 
         self._normal = normal / np.linalg.norm(normal)
         self._scalar = float(scalar)
@@ -186,7 +186,7 @@ class Plane(object):
     @normal.setter
     def normal(self, value):
         """Set the plane's normal"""
-        value = np.asarray(value, dtype=np.float)
+        value = np.asarray(value, dtype=np.float64)
         self._normal = value / np.linalg.norm(value)
 
     @property
@@ -409,8 +409,8 @@ class Mesh(object):
         assert np.array(vertices).shape[1] == 3
         assert np.array(faces).shape[1] == 4
 
-        self._vertices = np.array(vertices, dtype=np.float)
-        self._faces = np.array(faces, dtype=np.int)
+        self._vertices = np.array(vertices, dtype=np.float64)
+        self._faces = np.array(faces, dtype=np.int32)
         self._id = next(self._ids)
 
         if not name:
@@ -626,14 +626,14 @@ class Mesh(object):
 
     @vertices.setter
     def vertices(self, value):
-        self._vertices = np.asarray(value, dtype=np.float).copy()
+        self._vertices = np.asarray(value, dtype=np.float64).copy()
         # self._vertices.setflags(write=False)
         self.__internals__.clear()
         return
 
     @faces.setter
     def faces(self, value):
-        self._faces = np.asarray(value, dtype=np.int).copy()
+        self._faces = np.asarray(value, dtype=np.int32).copy()
         # self._faces.setflags(write=False)
         self.__internals__.clear()
         return
@@ -649,9 +649,9 @@ class Mesh(object):
         # quads_mask = np.invert(triangle_mask)
         # nb_quads = nf - nb_triangles
 
-        faces_areas = np.zeros(nf, dtype=np.float)
-        faces_normals = np.zeros((nf, 3), dtype=np.float)
-        faces_centers = np.zeros((nf, 3), dtype=np.float)
+        faces_areas = np.zeros(nf, dtype=np.float64)
+        faces_normals = np.zeros((nf, 3), dtype=np.float64)
+        faces_centers = np.zeros((nf, 3), dtype=np.float64)
 
         # Collectively dealing with triangles
         # triangles = _faces[triangle_mask]
@@ -1257,7 +1257,7 @@ class Mesh(object):
         # TODO: docstring
         # FIXME : code en doublon par rapport a la fonction _rodrigues du debut de module
 
-        angles = np.asarray(angles, dtype=np.float)
+        angles = np.asarray(angles, dtype=np.float64)
         theta = np.linalg.norm(angles)
         if theta == 0.:
             return np.eye(3)
@@ -1917,7 +1917,7 @@ class Mesh(object):
     def _compute_faces_integrals(self, sum_faces_contrib=False):  # TODO: implementer le sum_surface_contrib
 
         # TODO: Utiliser sum_faces_contrib
-        surface_integrals = np.zeros((15, self.nb_faces), dtype=np.float)
+        surface_integrals = np.zeros((15, self.nb_faces), dtype=np.float64)
 
         # First triangles
         if self.nb_triangles > 0:
@@ -2050,7 +2050,7 @@ class Mesh(object):
 
         s0, s1, s2, s3, s4, s5, s6, s7, s8 = self.get_surface_integrals()[:9].sum(axis=1)
 
-        cog = np.array([s0, s1, s2], dtype=np.float) / mass
+        cog = np.array([s0, s1, s2], dtype=np.float64) / mass
 
         xx = surf_density * (s7 + s8)
         yy = surf_density * (s6 + s8)
@@ -2064,7 +2064,7 @@ class Mesh(object):
     def _edges_stats(self):
         """Computes the min, max, and mean of the mesh's edge length"""
         vertices = self.vertices[self.faces]
-        edge_length = np.zeros((self.nb_faces, 4), dtype=np.float)
+        edge_length = np.zeros((self.nb_faces, 4), dtype=np.float64)
         for i in range(4):
             edge = vertices[:, i, :] - vertices[:, i - 1, :]
             edge_length[:, i] = np.sqrt(np.einsum('ij, ij -> i', edge, edge))
@@ -2102,7 +2102,7 @@ class Mesh(object):
         Explicit the integrals
         """
 
-        s_int = np.zeros((15, triangles_vertices.shape[0]), dtype=np.float)
+        s_int = np.zeros((15, triangles_vertices.shape[0]), dtype=np.float64)
 
         point_0, point_1, point_2 = list(map(_3DPointsArray, np.rollaxis(triangles_vertices, 1, 0)))
